% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_filedirs}
\alias{create_filedirs}
\title{Create file directories for data}
\usage{
create_filedirs(dir_to_save, id_variable)
}
\arguments{
\item{dir_to_save}{(char) path to save data onto}

\item{id_variable}{(char) name of variable with identifier of a site}
}
\value{
None
}
\description{
This script should only be run once to generate a separate directory per each site.
}
\details{
--- dir_to_save /
--- site_A /
--- site_B /
}
\examples{
create_filedirs('/my/new/folder/', 'var_of_interest')
}
