% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dplace}
\alias{dplace}
\title{A D-place Dataset}
\format{
\subsection{\code{dplace}}{

A data frame with 1988 rows and 5 variables:
\describe{
\item{SiteCode}{site identifier}
\item{lon}{longitude}
\item{lat}{latitude}
\item{code}{internal code for a d-place site}
\item{SubCont}{Sub-continent}
}
}
}
\source{
\url{https://raw.githubusercontent.com/D-PLACE/dplace-data/master/geo/societies_tdwg.json}
}
\usage{
dplace
}
\description{
A dataset with sites from the D-Place Project (https://d-place.org/).
}
\keyword{datasets}
