% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_raster}
\alias{extract_raster}
\title{Extract Raster}
\usage{
extract_raster(rast, clip_vec, func_name = mean)
}
\arguments{
\item{rast}{\itemize{
\item raster object (use terra)
}}

\item{clip_vec}{\itemize{
\item vector (typically an sf data.frame)
}}

\item{func_name}{\itemize{
\item function to summarize extracted values (default=MEAN)
}}
}
\value{
dataframe with summarized values
}
\description{
Extract Raster
}
\examples{
pts = make_point("my_site", 45, 45)
pts_buff = make_buffer(pts, 5000)
fpath = system.file("extdata", "africa_cropland_netgain.tif", package="endow")
r = terra::rast(fpath)
e = extract_raster(r, pts_buff)
}
