% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_filedirs}
\alias{generate_filedirs}
\title{Generate directories for data}
\usage{
generate_filedirs(dir_to_save, site_id, var_name)
}
\arguments{
\item{dir_to_save}{(char) path to save data onto}

\item{site_id}{(char) name of variable with identifier of a site}

\item{var_name}{(char) variable name (aka souce, e.g. 'night_light')}
}
\value{
dir_name (directory path)
}
\description{
This script should only be run once to generate a separate directory per each site.
}
\details{
--- dir_to_save /
--- site_A /
--- site_B /
}
\examples{
generate_filedirs('/my/new/folder/', site_id='AH', var_name='var_of_interest')
}
