% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_to_cell.R
\name{get_dist_to_cell}
\alias{get_dist_to_cell}
\title{Calculates distance to population center}
\usage{
get_dist_to_cell(celldata, path_to_save, k = 10, stat = "mean", ...)
}
\arguments{
\item{celldata}{(data frame) a dataframe downloaded from opencellid}

\item{path_to_save}{(char) a path to save procesed data}

\item{k}{(int) k for the number of neighbors}

\item{stat}{(char) either 'mean' or 'min', which knn dist to return}

\item{...}{}
}
\value{
None. Saves files to the 'processed' directory.
}
\description{
Calculates distance to population center
}
\examples{
# Kumasi coordinates: 6.695016860965001, -1.6179580414855728
pp = system.file('extdata', 'world_cities.csv', package='endow')

f = get_dist_to_pop(pp, '/my_folder/',
 site_id='KU',
 lon=-1.62, lat=6.7,
 var_name='cropland')

}
