% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_to_pop.R
\name{get_dist_to_pop}
\alias{get_dist_to_pop}
\title{Calculates distance to population center}
\usage{
get_dist_to_pop(cities_path, path_to_save, ...)
}
\arguments{
\item{cities_path}{(char) a path to file with cities}

\item{path_to_save}{(char) a path to save procesed data}

\item{...}{}
}
\value{
None. Saves files to the 'processed' directory.
}
\description{
Calculates distance to population center
}
\examples{
# Kumasi coordinates: 6.695016860965001, -1.6179580414855728
pp = system.file('extdata', 'world_cities.csv', package='endow')

f = get_dist_to_pop(pp, '/my_folder/', year=2017,
 year_var=as.POSIXct('2020-01-01 14:45:18',
 format="\%Y-\%m-\%d \%H:\%M:\%S",tz="UTC"),
 site_id='KU',
 FUN='sum',
 lon=-1.62, lat=6.7, dist=6000, var_name='cropland')

}
