% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_road_length.R
\name{get_osm_road_length}
\alias{get_osm_road_length}
\title{Calculates distance to population center}
\usage{
get_osm_road_length(path_to_save, ...)
}
\arguments{
\item{path_to_save}{(char) a path to save procesed data}

\item{...}{}
}
\value{
None. Saves files to the 'processed' directory.
}
\description{
Calculates distance to population center
}
\examples{
# Kumasi coordinates: 6.695016860965001, -1.6179580414855728
pp = system.file('extdata', 'world_cities.csv', package='endow')

f = get_dist_to_pop(pp, '/my_folder/',
 site_id='KU',
 lon=-1.62, lat=6.7,
 var_name='cropland')

}
