% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_buffer}
\alias{make_buffer}
\title{Make Buffer}
\usage{
make_buffer(sf_obj, dist = 5000)
}
\arguments{
\item{sf_obj}{a simple feature dataframe}

\item{dist}{(int) distance in meters to draw a buffer around}
}
\value{
buffer of an sf_obj
}
\description{
Make Buffer
}
\examples{
pts = make_point("my_site", 45, 45)
pts_buff = make_buffer(pts, 5000)
}
