% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_main_csv.R
\name{make_main_csv}
\alias{make_main_csv}
\title{make_main_csv}
\usage{
make_main_csv(processed_dir, output_dir)
}
\arguments{
\item{processed_dir}{(char) - dir with processed csv files}

\item{output_dir}{(char) - output dir where main.csv will be saved}
}
\value{
nothing
}
\description{
Collect single csv for each variable, year and buffer into
one big csv file for further processing.
}
