% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_point}
\alias{make_point}
\title{Make Point}
\usage{
make_point(site_id, lon, lat)
}
\arguments{
\item{site_id}{(id variable associated with a geographic point)}

\item{lon}{(longitude in decimal degrees)}

\item{lat}{(latitude in decimal degrees)}
}
\value{
a simple geometry feature
}
\description{
Make Point
}
\examples{
x = 45
y = 45
my_id = 'MM'
make_point(my_id, x, y)
}
