% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf_collector.R
\name{ncdf_collector}
\alias{ncdf_collector}
\title{NCDF collector}
\usage{
ncdf_collector(ncdf_path, path_to_save, year = NULL, year_var = NULL, ...)
}
\arguments{
\item{ncdf_path}{\itemize{
\item path to ncdf file
}}

\item{path_to_save}{\itemize{
\item save directory
}}

\item{year}{\itemize{
\item year to process the data for
}}

\item{year_var}{\itemize{
\item year with wave 1 start date
}}

\item{...}{\itemize{
\item any other params
}}
}
\value{
Writes files to disk. No return value.
}
\description{
NCDF collector
}
\examples{
ncdf_path = system.file("extdata", "world_soil_moisture_jan2016.nc",
 package="endow")
f = ncdf_collector(ncdf_path, '/my_folder/', year=2016,
 year_var=as.POSIXct('2020-01-01 14:45:18',
 format="\%Y-\%m-\%d \%H:\%M:\%S",tz="UTC"),
 site_id='KU', lon=-1.62, lat=6.7, dist=30000, var_name='soil_moisture')
}
