% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ncdf_to_tif}
\alias{ncdf_to_tif}
\title{NCDF to TIFF}
\usage{
ncdf_to_tif(ncdf_path, attr_name, year, fname)
}
\arguments{
\item{ncdf_path}{(char) path to ncdf file}

\item{attr_name}{(char) attribute name in ncdf file}

\item{year}{(int) year for which the data will be processed}

\item{fname}{(char) filename}
}
\value{
nothing
}
\description{
Converts ncdf files to tiff
}
\examples{

ncdf_path = system.file("extdata", "world_soil_moisture_jan2016.nc",
 package="endow")

ncdf_to_tif(ncdf_path=ncdf_path, attr_name = 'sm', year=2016,
 fname='soil_moisture')

}
