% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/night_collector.R
\name{night_collector}
\alias{night_collector}
\title{Collector}
\usage{
night_collector(raster_path, path_to_save, ...)
}
\arguments{
\item{raster_path}{(char) a path to an input raster to be processed}

\item{path_to_save}{(char) a directory to store the processed rasters}

\item{...}{other arguments from endow.utils functions}
}
\value{
None. Saves clipped rasters and processed csv files to /processed.
}
\description{
The function collects the data give an input raster and a coordinate.
}
\examples{
# Kumasi coordinates: 6.695016860965001, -1.6179580414855728
rast_path = system.file("extdata", "africa_cropland_netgain.tif", package="endow")
f = night_collector(rast_path, '/my_folder/', site_id='KU', lon=-1.62, lat=6.7, dist=6000, var_name='cropland')
}
